Use only one method for zchop. Slight change in symantics
Before I had one method for zchop! and many for zchop. Now, this is
reversed. The symantics is slightly changed, as reflected in the tests.
For the vast majority of cases, there is no difference. (Of course, some
person may only use the corner cases!)

I could probably do more tests for the new symantics.